requirements.txt
tensorflow==2.10.0
tensorflow-gpu==2.10.0
torch>=2.1.0
torchvision>=0.16.0
opencv-python>=4.0.0
numpy>=1.19.0
pandas>=1.0.0
scikit-learn>=0.24.0
Pillow>=8.0.0
matplotlib>=3.3.0
tqdm>=4.65.0 